+++ s3-source-connector/src/main/java/io/aiven/kafka/connect/s3/source/utils/FileReader.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2024 Aiven Oy
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.aiven.kafka.connect.s3.source.utils;
+
+import static io.aiven.kafka.connect.s3.source.config.S3SourceConfig.FETCH_PAGE_SIZE;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+
+import io.aiven.kafka.connect.s3.source.config.S3SourceConfig;
+
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.ListObjectsRequest;
+import com.amazonaws.services.s3.model.ObjectListing;
+import com.amazonaws.services.s3.model.S3Object;
+import com.amazonaws.services.s3.model.S3ObjectSummary;
+
+public class FileReader {
+
+    private final S3SourceConfig s3SourceConfig;
+    private final String bucketName;
+
+    public FileReader(final S3SourceConfig s3SourceConfig, final String bucketName) {
+        this.s3SourceConfig = s3SourceConfig;
+        this.bucketName = bucketName;
+    }
+
+    List<S3ObjectSummary> fetchObjectSummaries(final AmazonS3 s3Client) throws IOException {
+        final ObjectListing objectListing = s3Client.listObjects(new ListObjectsRequest().withBucketName(bucketName)
+                .withMaxKeys(s3SourceConfig.getInt(FETCH_PAGE_SIZE) * 2));
+
+        return new ArrayList<>(objectListing.getObjectSummaries());
+    }
+
+    InputStream getContent(final S3Object object) {
+        return object.getObjectContent();
+    }
+
+}
